refactor(step1): 디렉토리 변경
